backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
google-genai
Pillow